requirements.txt
py-cord==2.4.1
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2